pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xev-whisprflow"
version = "0.1.0"
description = "Push-to-talk voice dictation for Linux"
requires-python = ">=3.11"
dependencies = [
    "evdev>=1.6.1",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "deepgram-sdk>=3.0.0,<4.0.0",
    "groq>=0.9.0",
    "toml>=0.10.2",
    "python-dotenv>=1.0.0",
    "PyGObject>=3.44.0",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "python-multipart>=0.0.9",
    "pystray>=0.19.5",
    "Pillow>=10.0.0",
]

[project.scripts]
xev-whisprflow = "xev_whisprflow.main:cli_entry"

[tool.setuptools.packages.find]
where = ["."]
include = ["xev_whisprflow*"]
